// src/app/main/pages/compartidos/interfaces/ImodeloPromedio.ts
export interface ImodeloPromedio {
    mprCodigo?: number;
    reanleCodigo?: number;
    tipeduCodigo?: number;
    graCodigo?: number;
    mprDenominacion?: string;
    mprDesde?: number;
    mprHasta?: number;
    mprEstado?: number;
    mprNemonico?: string;

    tipoEducacion?: any;
    grado?: any;
}